requirements.txt
setuptools
wheel
requests
selenium
locustio
Faker